requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
idna==3.7
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.8
